# gabarita-ai-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Gabarita AI API
  description: |
    API do sistema Gabarita AI para geração de questões, gerenciamento de usuários e planos.
    
    ## Autenticação
    A API utiliza autenticação via JWT tokens. Inclua o token no header:
    ```
    Authorization: Bearer <seu_token_jwt>
    ```
    
    ## Formato de Resposta
    Todas as respostas seguem o padrão:
    ```json
    {
      "success": true|false,
      "data": { /* dados da resposta */ },
      "message": "Mensagem descritiva",
      "error": "Tipo do erro (apenas em caso de erro)"
    }
    ```
  version: 1.0.0
  contact:
    name: Gabarita AI Support
    email: suporte@gabaritai.app.br
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gabarita-ai-backend.onrender.com
    description: Servidor de Produção
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento
paths:
...
  /api/usuarios/{user_id}:
    get:
      tags:
        - Usuários (Aliases)
      summary: Obter usuário (Alias)
      description: Alias para /api/user/{user_id}
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        $ref: '#/paths/~1api~1user~1{user_id}/get/responses'

    put:
      tags:
        - Usuários (Aliases)
      summary: Atualizar usuário (Alias)
      description: Alias para /api/user/{user_id}
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/paths/~1api~1user~1{user_id}/put/requestBody'
      responses:
        $ref: '#/paths/~1api~1user~1{user_id}/put/responses'
  /api/usuarios/perfil:
    get:
      tags:
        - Usuários (Aliases)
      summary: Obter perfil (Alias)
      description: Alias para /api/user/profile
      security:
        - bearerAuth: []
      responses:
        $ref: '#/paths/~1api~1user~1profile/get/responses'

    put:
      tags:
        - Usuários (Aliases)
      summary: Atualizar perfil (Alias)
      description: Alias para /api/user/profile
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/paths/~1api~1user~1profile/put/requestBody'
      responses:
        $ref: '#/paths/~1api~1user~1profile/put/responses'

  # OPÇÕES
...
  /api/questoes/gerar:
    post:
      tags:
        - Questões
      summary: Gerar nova questão
      description: Gera uma questão personalizada baseada nos parâmetros fornecidos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cargo
                - bloco
                - conteudo
              properties:
                cargo:
                  type: string
                  description: Cargo para o qual gerar a questão
                bloco:
                  type: string
                  description: Bloco de conteúdo
                conteudo:
                  type: string
                  description: Conteúdo específico
                dificuldade:
                  type: string
                  enum: [facil, medio, dificil]
                  default: medio
      responses:
        '200':
          description: Questão gerada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Question'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operação realizada com sucesso
      required:
        - success
        - message
...
    Question:
      type: object
      properties:
        id:
          type: string
          description: ID único da questão
        enunciado:
          type: string
          description: Texto da questão
        alternativas:
          type: object
          properties:
            A:
              type: string
            B:
              type: string
            C:
              type: string
            D:
              type: string
            E:
              type: string
          description: Alternativas da questão
        cargo:
          type: string
          description: Cargo relacionado
        bloco:
          type: string
          description: Bloco de conteúdo
        conteudo:
          type: string
          description: Conteúdo específico
        dificuldade:
          type: string
          enum: [facil, medio, dificil]
          description: Nível de dificuldade
      required:
        - id
        - enunciado
        - alternativas
        - cargo
        - bloco
        - conteudo
        - dificuldade
...